refactor: 포인트 관련 필드 타입 BigDecimal로 변경

+++ src/main/java/com/bootcamp/paymentdemo/point/repository/PointRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -15,11 +16,11 @@ public interface PointRepository extends JpaRepository<PointTransaction, Long> {
     List<PointTransaction> findByUser_UserIdOrderByCreatedAtDesc(Long userId);
 
     @Query("SELECT SUM(p.amount) FROM PointTransaction p WHERE p.user.userId = :userId")
-    Long calculatePointBalance(@Param("userId") Long userId);
+    BigDecimal calculatePointBalance(@Param("userId") Long userId);
 
     Optional<PointTransaction> findByOrderIdAndType(Long orderId, PointType pointType);
 
-    List<PointTransaction> findByUser_UserIdAndTypeAndRemainingAmountGreaterThanAndExpiresAtAfterOrderByExpiresAtAsc(Long userId, PointType pointType, int i, LocalDate now);
+    List<PointTransaction> findByUser_UserIdAndTypeAndRemainingAmountGreaterThanAndExpiresAtAfterOrderByExpiresAtAsc(Long userId, PointType pointType, BigDecimal i, LocalDate now);
 
-    List<PointTransaction> findByTypeAndRemainingAmountGreaterThanAndExpiresAtBefore(PointType pointType, int i, LocalDate now);
+    List<PointTransaction> findByTypeAndRemainingAmountGreaterThanAndExpiresAtBefore(PointType pointType, BigDecimal i, LocalDate now);
 }
